==> R/03_augment.qmd <==
---
title: "03_aug"
format: html
editor: visual
---

# 03_augment

### Prepping

Clean Work Environment and Load Packages

```{r}
#| message: false

rm(list = ls())
library(tidyverse)

```

### Load Data

```{r}
#| message: false

data_path <-  '~/projects/Group25_Project/data'

data_barra <- read_rds(file = file.path(data_path,
                                        '02_dat_clean_barracoda.RDS'))
data_flow <- read_rds(file = file.path(data_path,
                                       '02_dat_clean_flowjo.RDS'))

outfile_joined_path <- file.path(data_path,
                                 '03_dat_aug_joined.RDS')
outfile_sig_pairing_path <- file.path(data_path,
                                      '03_dat_aug_sig_pairing.RDS')

```

### Join Barracoda and Flowjo data

Barracoda and flowjo data are joined into one tibble. They do not overlap perfectly, therfore the innerjoin causes some dataloss from both sources.

Then the estimated frequency is calculated as such: $Est.Frequency = (count_{pMHC}/count_{total}) * fraction_{PE positive}$

```{r}

joined_data <- data_barra |>
              inner_join(data_flow, 
                         by=c('sample_ID',
                              'KIR_blocking')) |>
              group_by(sample_ID) |> 
              mutate(total_count = (sum(count,
                                        na.rm = TRUE)),
                     count_fraction = (count / total_count), # count fraction
                     est_frequency = count_fraction*PE_positive) |>
              ungroup()
                
```

### Filtering for significant data

We decided that the only significant data was data with a p-value \< 0.001 and a log fold change \> 2.

**NB:**

-   if we filter the significant before the pairing we get 22 significant observations, meaning 11 pairs

-   if we filter the significant after the pairing we get 41 significant observations, some of which are no longer in pairs

This is why the filtering happens at this step and not later.

```{r}

sig_data <- joined_data |> 
  filter(p < 0.001 & log_fold_change > 2)

```

### **Adding a "paired" column**

This column is equal to 1 when there is another row with the same sample_ID, cohort, HLA type and peptide, that has the opposite KIR blocking status (blocked = 1, non-blocked = 0).

This means that the same pMHC multimer has been tested in both experiments, i.e. we have paired data.

```{r}

joined_data <- joined_data |> 
  group_by(sample_ID, cohort, HLA, peptide) |>
  mutate(paired = factor(
                  if_else((all(c(0, 1) %in% KIR_blocking)), 1, 0)),
         .after = cohort) |>
  ungroup()

sig_data <- sig_data |> 
            group_by(sample_ID, cohort, HLA, peptide) |>
            mutate(paired = factor(
                            if_else((all(c(0, 1) %in% KIR_blocking)), 1, 0)),
                   .after = cohort) |>
            ungroup()

# just for observation, not saved
paired_data <- sig_data |> 
  filter(paired == 1)

```

We go from 31796 obs to 22 obs (meaning 11 pairs)

Barracoda: 31796 obs (+ 56 Flowjo obs)

Joined: 22819 obs

Significant: 542 obs

Paired: 22 obs

### **S**ave to csv

```{r}

# the joined data, not filtered for significance, but pairs are labelled
saveRDS(joined_data,
        file = outfile_joined_path)

# only significant data, pairs are labelled
saveRDS(sig_data,
        file = outfile_sig_pairing_path)
```
