==> R/06_analysis2.qmd <==
---
title: "Group 25: Project Report 2023"
format: html
editor: visual
---

# 06_analysis2

### Prepping

Clean Work Environment and Load Packages

```{r}
#| message: false

rm(list = ls())
library(tidyverse)
library(RColorBrewer)
```

### Loading data

```{r}
#| message: false

data_path <-  '~/projects/Group25_Project/data'
data_joined <- read_rds(file = file.path(data_path,
                                         '03_dat_aug_joined.RDS'))
data_sig_pairing <- read_rds(file = file.path(data_path,
                                              
                                              '03_dat_aug_sig_pairing.RDS'))
```

### Defining some settings once

```{r}
#| message: false

# alpha, size and colors for volcano plots
a <- 0.25                                # alpha
s <- 0.005                               # size

colors <- brewer.pal(9, "Paired")[c(2, 6)]
volcano_colors  <- c(colors[1], colors[2], 'grey')

# adding a response level
data_joined <- data_joined |>
  mutate(response =
           factor(
             case_when(log_fold_change > 2 & p < 0.001 ~ "Positive",
                           log_fold_change < -2 & p <= 0.001 ~ "Negative",
                           TRUE ~ "Insignificant"),
             levels = c("Positive", "Negative", "Insignificant")))
```

### Volcano plots, grouped by cohort and HLA

```{r}
#| message: false

data_joined |>
  group_by(HLA, cohort) |>
  ggplot(aes(x = log_fold_change,
             y = - log10(p),
             color = response)) +
  theme_bw() +
  geom_vline(xintercept = 2,
             linetype = "dashed",
             color = "grey") +
  geom_vline(xintercept = -2,
             linetype = "dashed",
             color = "grey") +
  geom_vline(xintercept = 0,
             linetype = "solid",
             color = "grey") +
  geom_hline(yintercept = -log10(0.001),
             linetype = "dashed",
             color = "grey") +
  geom_point(alpha = a, size = s) +
  labs(x = "Fold change (log2)",
       y = "P-value (-log10)") +
  scale_color_manual(values = volcano_colors) +
    labs(x = "Fold change (log2)",
       y = "P-value (-log10)",
       color = "Response",
       title = "Response levels accross all cohorts, and HLA types",
       subtitle = "For both paired and unpaired data") +
  guides(color = guide_legend(override.aes = list(size=2,
                                                  alpha = 0.75))) +
  theme(axis.title.y = element_text(margin = margin(t = 0,
                                                    r = 10,
                                                    b = 0,
                                                    l = 0)),
        axis.title.x = element_text(margin = margin(t = 10,
                                                    r = 0,
                                                    b = 0,
                                                    l = 0))) +
  facet_grid(HLA ~ cohort)

ggsave("../results/Volcano_cohorts_HLA.png", 
       bg = 'white')
```

### Volcano plots all cohorts

```{r}
#| message: false

data_joined |>
  group_by(HLA, cohort) |>
  ggplot(aes(x = log_fold_change,
             y = - log10(p),
             color = response)) +
  theme_bw() +
  geom_vline(xintercept = 2,
             linetype = "dashed",
             color = "grey") +
  geom_vline(xintercept = -2,
             linetype = "dashed",
             color = "grey") +
  geom_vline(xintercept = 0,
             linetype = "solid",
             color = "grey") +
  geom_hline(yintercept = -log10(0.001),
             linetype = "dashed",
             color = "grey") +
  geom_point(alpha = a, size = s) +
  labs(x = "Fold change (log2)",
       y = "P-value (-log10)") +
  scale_color_manual(values = volcano_colors) +
    labs(x = "Fold change (log2)",
       y = "P-value (-log10)",
       color = "Response",
       title = "Response levels accross all cohorts, both paired and unpaired data",
       subtitle = "AP = , BS =, BC = ") +
  guides(color = guide_legend(override.aes = list(size=2,
                                                  alpha = 0.75))) +
  theme(axis.title.y = element_text(margin = margin(t = 0,
                                                    r = 10,
                                                    b = 0,
                                                    l = 0)),
        axis.title.x = element_text(margin = margin(t = 10,
                                                    r = 0,
                                                    b = 0,
                                                    l = 0)),
        legend.position = c(0.78, 0.25)) +
  facet_wrap(~ cohort, 
             nrow = 2)

ggsave("../results/Volcano_cohorts.png", 
       bg = 'white')
```

### Volcano plot for all data

```{r}
#| message: false

data_joined |>
  ggplot(aes(x = log_fold_change,
             y = - log10(p),
             color = response)) +
  theme_bw() +
  geom_vline(xintercept = 2,
             linetype = "dashed",
             color = "grey") +
  geom_vline(xintercept = -2,
             linetype = "dashed",
             color = "grey") +
  geom_vline(xintercept = 0,
             linetype = "solid",
             color = "grey") +
  geom_hline(yintercept = -log10(0.001),
             linetype = "dashed",
             color = "grey") +
  geom_point(alpha = a, size = s) +
  scale_color_manual(values = volcano_colors) +
  labs(x = "Fold change (log2)",
       y = "P-value (-log10)",
       color = "Response",
       title = "Response levels accross whole dataset",
       subtitle = "Only small subset (n = 542) is significant at p < 0.001 and LFC >= 2") +
  guides(color = guide_legend(override.aes = list(size=2,
                                                  alpha = 0.75))) +
  theme(axis.title.y = element_text(margin = margin(t = 0,
                                                    r = 10,
                                                    b = 0,
                                                    l = 0)),
        axis.title.x = element_text(margin = margin(t = 10,
                                                    r = 0,
                                                    b = 0,
                                                    l = 0)))

ggsave("../results/Volcano_all.png", 
       bg = 'white')
```
